fix: flakey gen box test
